Add hash validation functionality
New field variables have been added to store file paths used for hash validation. Plus, the code for generating and comparing the hash values of two files is now included in the 'ValidateHashButtonClick' method. Additionally, the user interface has been updated with a new 'Validate Hash' button and a textbox to display the computed hash.